Move the AzureTokenService.java to roleClient from init
