--- init-service/src/main/java/uk/ac/ox/ndph/mts/init_service/config/package-info.java
@@ -1,4 +1,0 @@
-/**
- * The trial config service package
- */
-package uk.ac.ox.ndph.mts.init_service.config;
