+++ init-service/src/test/java/uk/ac/ox/ndph/mts/init_service/service/SiteServiceTest.java
@@ -12,10 +12,10 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.web.reactive.function.client.WebClient;
-import uk.ac.ox.ndph.mts.init_service.config.AzureTokenService;
 import uk.ac.ox.ndph.mts.init_service.exception.DependentServiceException;
 import uk.ac.ox.ndph.mts.init_service.model.IDResponse;
 import uk.ac.ox.ndph.mts.init_service.model.Site;
+import uk.ac.ox.ndph.mts.roleserviceclient.configuration.AzureTokenService;
 import uk.ac.ox.ndph.mts.roleserviceclient.configuration.WebClientConfig;
 
 import java.io.IOException;
